pyproject.toml
[project]
name = "agent-trading-company"
version = "0.1.0"
description = "Multi-agent trading system for KR/US using KIS."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "feedparser>=6.0",
  "openai>=1.12",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
